Set guest_os_type to "Ubuntu_64" for VirtualBox packer build
When building on OS X, the VirtualBox build would get stuck with
the message:

“This kernel requires an x86-64 CPU, but only detected an i686 CPU.”

In the Packer docs for virtualbox-iso, it says:

https://www.packer.io/docs/builders/virtualbox-iso.html#guest_os_type

guest_os_type (string) - The guest OS type being installed. By default this is other, but you can get dramatic performance improvements by setting this to the proper value. To view all available values for this run VBoxManage list ostypes. Setting the correct value hints to VirtualBox how to optimize the virtual hardware to work best with that operating system.

Setting this value to Ubuntu_64 fixes the build problem.